Test with fewer Matcher tests
Narrow in on Appveyor failure on Python 2.7
